requirements.txt
Flask
torch
torchvision
torchxrayvision
pydicom
Pillow
numpy
matplotlib
